+++ mayan/apps/documents/events.py
@@ -9,9 +9,6 @@ namespace = EventTypeNamespace(label=_('Documents'), name='documents')
 event_document_create = namespace.add_event_type(
     label=_('Document created'), name='document_create'
 )
-event_document_download = namespace.add_event_type(
-    label=_('Document downloaded'), name='document_download'
-)
 event_document_properties_edit = namespace.add_event_type(
     label=_('Document properties edited'), name='document_edit'
 )
@@ -22,13 +19,16 @@ event_document_viewed = namespace.add_event_type(
 # Document File
 
 event_document_file_deleted = namespace.add_event_type(
-    label=_('Document file deleted'), name='document_version_revert'
+    label=_('Document file deleted'), name='document_file_deleted'
+)
+event_document_file_downloaded = namespace.add_event_type(
+    label=_('Document file downloaded'), name='document_file_downloaded'
 )
 event_document_file_new = namespace.add_event_type(
-    label=_('Document file uploaded'), name='document_new_version'
+    label=_('Document file uploaded'), name='document_file_new'
 )
 event_document_file_pre_save = namespace.add_event_type(
-    label=_('Document file created'), name='document_version_pre_save'
+    label=_('Document file created'), name='document_file_pre_save'
 )
 
 # Document type
@@ -60,4 +60,16 @@ event_document_version_edited = namespace.add_event_type(
 
 event_document_trashed = namespace.add_event_type(
     label=_('Document trashed'), name='document_trashed'
+)
+
+# Historic events
+
+event_document_download = namespace.add_event_type(
+    label=_('Document downloaded'), name='document_download'
+)
+event_document_version_revert = namespace.add_event_type(
+    label=_('Document version revert'), name='document_version_revert'
+)
+event_document_version_new = namespace.add_event_type(
+    label=_('Document version uploaded'), name='document_new_version'
 )